// worker-service/src/governance/dto/gov-action-proposal.dto.ts
import { Timestamp } from 'typeorm';
import { Vote } from '../entities/vote.entity';

export class GovActionProposalDto {
  id: string;
  votingAnchorId: string;
  title: string;
  abstract: string;
  govActionType: string;
  govMetadataUrl: string;
  status: string;
  endTime: Date;
  txHash: string;
  submitTime: Date;
  votes: Vote[];
  createdAt: Timestamp;
  updatedAt: Timestamp;
}